apps/transcribe/pyproject.toml
[project]
name = "transcribe"
version = "0.1.0"
description = "Audio recording and transcription with Whisper"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-whisper>=20250625",
    "streamlit>=1.40.0",
]
